refactored the files for /form and /form/customize
