+++ src/components/FormCustomize/FormCustomize.jsx
@@ -1,21 +1,17 @@
+import { useNavigate } from "react-router-dom";
+import { useLocation } from "react-router-dom";
 import { useEffect, useState } from "react";
-import TokenGallery from '../TokenGallery/TokenGallery';
-import GenderSelection from "../GenderSelection/GenderSelection";
+
 
 function FormCustomize() {
-    const [selectedToken, setSelectedToken] = useState(null)
-    const [selectedGender, setSelectedGender] = useState(null)
-    const [availableStats, setAvailableStats] = useState([])
-    const [selectedStats, setSelectedStats] = useState({})
-    const [statPointsLeft, setStatPointsLeft] = useState(10)
-    const BASE_STAT_VALUE = 8
-    const MAX_STAT_POINTS_PER_STAT = 5
-    const [confirmedStats, setConfirmedStats] = useState(false)
-    const [availableSkills, setAvailableSkills] = useState([])
-    const [selectedSkills, setSelectedSkills] = useState({})
-    const [skillPointsLeft, setSkillPointsLeft] = useState(5)
-    const MAX_POINTS_PER_SKILL = 3;
-    const [confirmedSkill, setConfirmedSkill] = useState(false)
+    const navigate = useNavigate()
+
+    const { state } = useLocation();    
+    const selectedToken = state?.selectedToken;
+    const selectedGender = state?.selectedGender;
+    const selectedStats = state?.selectedStats;
+    const selectedSkills = state?.selectedSkills;
+
     const [availableCantrips, setAvailableCantrips] = useState([])
     const [selectedCantrips, setSelectedCantrips] = useState([])
     const [confirmedCantrips, setConfirmedCantrips] = useState(false);
@@ -31,36 +27,7 @@ function FormCustomize() {
     const [confirmedWeapon, setConfirmedWeapon] = useState(false)
     const [selectedArmor, setSelectedArmor] = useState(null)
     const [confirmedArmor, setConfirmedArmor] = useState(false)
-
-    
-
-    useEffect(() => {
-        if (!selectedToken) return
-
-        fetch('/stats_data.json')
-        .then((response) => response.json())
-        .then((data) => {
-            const classStats = data
-            // console.log("Stats LOADED:", classStats)
-            setAvailableStats(classStats || [])
-        })
-    }, [selectedToken])
-
-    useEffect(() => {
-        if(!selectedToken) return
-
-        fetch('/skills_data.json')
-        .then((response) => response.json())
-        .then((data) => {
-            // console.log("Skills loaded:", data);
-            // console.log("Selected token class:", selectedToken.class);
-
-            const filteredSkills = data.filter((skill) => {
-                return skill.recommendedFor.includes(selectedToken.class)
-            })
-            setAvailableSkills(filteredSkills)
-        })
-    }, [selectedToken])
+    const [characterName, setCharacterName] = useState("")  
 
     useEffect(() => {
         if(!selectedToken) return
@@ -68,8 +35,6 @@ function FormCustomize() {
         fetch('/spells_n_cantrips_data.json')
         .then((response) => response.json())
         .then((data) => {
-            // console.log("Selected class:", selectedToken.class)
-            // console.log("Available cantrips before filter:", data.cantrips)
             const filterCantrips = data.cantrips.filter((cantrip) => {
                 return cantrip.recommendedFor.includes(selectedToken.class)
             })
@@ -79,8 +44,6 @@ function FormCustomize() {
             const filterLevel2Spells = data.spells_level_2.filter((level2) => {
                 return level2.recommendedFor.includes(selectedToken.class)
             })
-            // console.log("Level 1 Spells:", filterLevel1Spells)
-            // console.log("Level 2 Spells:", filterLevel2Spells)
             setAvailableCantrips(filterCantrips)
             setAvailableLevel1Spells(filterLevel1Spells)
             setAvailableLevel2Spells(filterLevel2Spells)
@@ -104,29 +67,9 @@ function FormCustomize() {
         })
     }, [selectedToken])
 
-
     return (
         <section>
-            <div className='form-selection'>{/* Token */}
-                <TokenGallery onSelect={setSelectedToken} />
-            </div>
-
-            {selectedToken && (
-                <div>
-                    <h3>Token Selected!</h3>
-                    <p>{selectedToken.race} {selectedToken.class}</p>
-                    <div className="token-card confirmed-token">
-                        <img src={selectedToken.url} alt="preview" style={{ width: '100px' }} />
-                    </div>
-                </div>
-            )}
-
-            {selectedToken && (
-                <div className='form-selection'>{/* Gender */}
-                    <GenderSelection onSelectGender={setSelectedGender} />
-                </div>
-            )}
-
+        
             {confirmedWeapon && (
                 <div className="confirmed-gear">
                     <p><strong>Weapon:</strong> {selectedWeapon}</p>
@@ -237,115 +180,8 @@ function FormCustomize() {
                     </button>
                 </div>
             )}
-
-            {selectedGender && (
-                <div className='form-selection'>{/* Stats */}
-                    {availableStats.map((stat) => {
-                        const assignedPoints = selectedStats[stat.name] || 0
-                        return (
-                            <div className="stat-card" key={stat.name}>
-                                <h4>{stat.name} - {BASE_STAT_VALUE + assignedPoints}</h4>
-                                <p>{stat.description}</p>
-                                {!confirmedStats && (
-                                  <p><strong>Recommended for:</strong> {stat.recommendedFor.join(", ")}</p>
-                                )}
-                            
-                                {!confirmedStats && (
-                                    <div className="stat-buttons">
-                                        {/* - Button */}
-                                        <button
-                                            onClick={() => {
-                                                if (assignedPoints > 0) {
-                                                    setSelectedStats(previousState => ({
-                                                        ...previousState,
-                                                        [stat.name]: assignedPoints - 1
-                                                    }))
-                                                    setStatPointsLeft(previousState => previousState + 1)
-                                                }
-                                            }}
-                                            disabled={assignedPoints === 0}
-                                        >–</button>
-
-                                        {/* + Button */}
-                                        <button
-                                            onClick={() => {
-                                                if (statPointsLeft > 0 && assignedPoints < MAX_STAT_POINTS_PER_STAT) {
-                                                    setSelectedStats(previousState => ({
-                                                        ...previousState,
-                                                        [stat.name]: assignedPoints + 1
-                                                    }))
-                                                    setStatPointsLeft(previousState => previousState - 1)
-                                                }
-                                            }}
-                                            disabled={statPointsLeft === 0 || assignedPoints >= MAX_STAT_POINTS_PER_STAT}
-                                        >+</button>
-                                    </div>
-                                )}
-                            </div>
-                        )
-                    })}
-
-                    <p>Points Remaining: {statPointsLeft}</p>
-                    <button
-                        className="confirm-button"
-                        onClick={() => setConfirmedStats(true)}
-                        disabled={statPointsLeft > 0}>
-                            Confirm Choices
-                    </button>
-                </div>
-            )}
-
-            {confirmedStats && (
-                <div className='form-selection'>{/* Skills */}
-                    {availableSkills.map(skill => {
-                        const assignedSkillPoints = selectedSkills[skill.name] || 0
-                        return (
-                            <div key={skill.name} className="skill-card">
-                                <h4>{skill.name} - {assignedSkillPoints}</h4>
-                                <p>{skill.description}</p>
-                                <div className="skill-buttons">
-                                    {/* - Button */}
-                                    <button
-                                            onClick={() => {
-                                                if (assignedSkillPoints > 0) {
-                                                    setSelectedSkills(previousState => ({
-                                                        ...previousState,
-                                                        [skill.name]: assignedSkillPoints - 1
-                                                    }))
-                                                    setSkillPointsLeft(previousState => previousState + 1)
-                                                }
-                                            }}
-                                            disabled={assignedSkillPoints === 0}
-                                        >–</button>
-
-                                        {/* + Button */}
-                                        <button
-                                            onClick={() => {
-                                                if (skillPointsLeft > 0 && assignedSkillPoints < MAX_POINTS_PER_SKILL) {
-                                                    setSelectedSkills(previousState => ({
-                                                        ...previousState,
-                                                        [skill.name]: assignedSkillPoints + 1
-                                                    }))
-                                                    setSkillPointsLeft(previousState => previousState - 1)
-                                                }
-                                            }}
-                                            disabled={skillPointsLeft === 0 || assignedSkillPoints >= MAX_POINTS_PER_SKILL}
-                                        >+</button>
-                                </div>
-                            </div>
-                        )
-                    })}
-                    <p>Points Remaining: {skillPointsLeft}</p>
-                    <button
-                        className="confirm-button"
-                        onClick={() => setConfirmedSkill(true)}
-                        disabled={skillPointsLeft > 0}>
-                            Confirm Choices
-                    </button>
-                </div>
-            )}
             
-            {confirmedSkill && !confirmedCantrips && (
+            {confirmedWeapon && confirmedArmor && !confirmedCantrips && (
                 <div className='form-selection'> {/* Cantrips */}
                     <h3>Select Your Cantrips, Hero</h3>
                         {availableCantrips.map((cantrip) => {
@@ -380,8 +216,6 @@ function FormCustomize() {
                     </button>
                 </div>
             )}
-
-
             
             {confirmedCantrips && !confirmedLevel1Spells && (
                 <div className='form-selection'> {/* Spells */}
@@ -419,7 +253,7 @@ function FormCustomize() {
                 </div>
             )}
 
-            {confirmedCantrips && !confirmedLevel2Spells && (
+            {confirmedLevel1Spells && !confirmedLevel2Spells && (
                 <div className='form-selection'> {/* Spells */}
                     <h3>Select Your Level 2 Spells, Hero</h3>
                         {availableLevel2Spells.map((spells_level_2) => {
@@ -455,13 +289,40 @@ function FormCustomize() {
                 </div>
             )}
 
-
-            <div className='form-selection'> {/* Name */}
-                {/* Add content for Name selection here */}
-            </div>
-            <div className='form-selection submit-button'> {/* Submit */}
-                {/* Add content for Submit button here */}
-            </div>
+            {confirmedLevel2Spells && (
+                <div className='form-selection'> {/* Name */}
+                    <h3>Name Your Hero</h3>
+                    <input 
+                        type="text"
+                        value={characterName}
+                        onChange={(event) => {
+                            setCharacterName(event.target.value)
+                        }}
+                        placeholder="Enter Character Name..."                        
+                    />
+                </div>
+            )}
+            
+            {characterName && (
+                <div className='form-selection'> {/* Submit Button */}
+                    <button
+                    className="confirm-button"
+                    onClick={() => {
+                        navigate("/sheet", {
+                            state: {
+                              skills: selectedSkills,
+                              cantrips: selectedCantrips,
+                              level1Spells: selectedLevel1Spells,
+                              level2Spells: selectedLevel2Spells,
+                              weapon: selectedWeapon,
+                              armor: selectedArmor
+                            }
+                        })
+                    }}
+                    >Submit Your Dweller!
+                    </button>
+                </div>
+            )}
         </section>
     )
 }
